refactoring methods

+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/jhompo/api/dtos/ApplicationTypeDTO.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,8 +15,8 @@ public class ApplicationTypeDTO {
 
     private Integer id;
     private String name;
-    private double minimum_amount;
-    private double maximum_amount;
+    private BigDecimal minimum_amount;
+    private BigDecimal maximum_amount;
     private double interest_rate;
     private boolean automatic_validation;
 }
